doc: clarify how to use env vars in nginx
nginx does not inherit env vars from the main proccess to its worker.
You must use the `env` keyword to specify which envs should be passed.

If this part is omitted, the plugin is loaded in the worker process
without envs and thus will fallback to the default values and may not
work.