feat: Add flushing destructor for Tracer
Rational: In short lived applications (e.g. Kubernetes Jobs) most
of the time all traces get lost becasue they are not synced. Users
need to call `Close` manually, but as all spans hold a shared_ptr
on the tracer it is hard to have the correct place to call it.
Calling it in the destructor assures all spans will be included in the
flush.